pyproject.toml
[project]
name = "triton-project"
version = "0.1.0"
description = ""
authors = [
    {name = "alex",email = "alex.tokmashov@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "numpy (>=2.4.1,<3.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "tritonclient (>=2.64.0,<3.0.0)",
    "gevent (>=25.9.1,<26.0.0)",
    "geventhttpclient (>=2.3.7,<3.0.0)",
    "python-multipart (>=0.0.21,<0.0.22)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "opencv-python (>=4.13.0.90,<5.0.0.0)",
    "onnx (>=1.20.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
